+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 - EntityComboBoxModel now an interface.
 - EntityComboBoxModel.ForeignKeyFilters renamed ForeignKeyFilter.
 - EntityComboBoxModel.ForeignKeyComboBoxModelFactory and ForeignKeyComboBoxModelLinker removed, ForeignKeyFilter.link() added.
+- EntityComboBoxModel.Builder added, related refactoring.
 ### is.codion.swing.framework.ui
 - EntityTablePanel.Config.cellRenderer() and cellEditor() added.
 - EntityComboBox.model() added for consistency.